수정/삭제 기능 보안 처리

+++ src/main/java/com/mugen/web/QuestionController.java
@@ -51,21 +51,50 @@ public class QuestionController {
 	}
 	
 	@GetMapping("/{id}/form")
-	public String updateForm(@PathVariable Long id, Model model) {
+	public String updateForm(@PathVariable Long id, Model model, HttpSession session) {
+		if(!HttpSessionUtils.isLoginUser(session)) {
+			return "redirect:/users/loginForm";
+		}
+		
+		User loginUser = HttpSessionUtils.getUserFromSession(session);
+		Question question = questionRepo.findOne(id);
+		if(!question.isSameWriter(loginUser)) {
+			return "redirect:/users/loginForm";
+		}
+				
 		model.addAttribute("question", questionRepo.findOne(id));
 		return "qna/updateForm";
 	}
 	
 	@PutMapping("/{id}")
-	public String update(@PathVariable Long id, String title, String contents) {
+	public String update(@PathVariable Long id, String title, String contents, HttpSession session) {
+		if(!HttpSessionUtils.isLoginUser(session)) {
+			return "redirect:/users/loginForm";
+		}
+		
+		User loginUser = HttpSessionUtils.getUserFromSession(session);
 		Question question = questionRepo.findOne(id);
+		if(!question.isSameWriter(loginUser)) {
+			return "redirect:/users/loginForm";
+		}
+		
 		question.update(title, contents);
 		questionRepo.save(question);
 		return String.format("redirect:/questions/%d", id);
 	}
 	
 	@DeleteMapping("/{id}")
-	public String delete(@PathVariable Long id) {
+	public String delete(@PathVariable Long id, HttpSession session) {
+		if(!HttpSessionUtils.isLoginUser(session)) {
+			return "redirect:/users/loginForm";
+		}
+		
+		User loginUser = HttpSessionUtils.getUserFromSession(session);
+		Question question = questionRepo.findOne(id);
+		if(!question.isSameWriter(loginUser)) {
+			return "redirect:/users/loginForm";
+		}
+		
 		questionRepo.delete(id);
 		return "redirect:/";
 	}
